v2 fix

+++ src/main/java/pro/dto/LimitDto.java
@@ -1,4 +1,4 @@
 package pro.dto;
 
-public record LimitDto(Long id, Float reservation) {
+public record LimitDto(Long id, Double reservation) {
 }
