+++ src/main/java/pro/exception/LimitException.java
@@ -4,10 +4,10 @@ import lombok.Getter;
 
 @Getter
 public class LimitException extends RuntimeException {
-    private final Float currentLimit;
-    private final Float reservation;
+    private final Double currentLimit;
+    private final Double reservation;
 
-    public LimitException(Float currentLimit, Float reservation) {
+    public LimitException(Double currentLimit, Double reservation) {
         this.currentLimit = currentLimit;
         this.reservation = reservation;
     }
